wereyouhere/requirements.txt
# for detecting sqlite databases
python-magic
